move request from data.py to request.py move response from data.py to response.py
