+++ python/tchannel/tornado/hyperbahn.py
@@ -21,11 +21,11 @@
 from __future__ import absolute_import
 
 import json
+
 import tornado.gen
 
 from ..messages.error import ErrorCode
 from .tchannel import TChannel
-
 
 DEFAULT_TTL = 60  # seconds
