+++ python/tchannel/tornado/request.py
@@ -1,0 +1,120 @@
+# Copyright (c) 2015 Uber Technologies, Inc.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+
+from __future__ import absolute_import
+
+import tornado
+import tornado.gen
+
+from ..messages.common import FlagsType
+from ..messages.common import StreamState
+from ..zipkin.trace import Trace
+from .util import get_arg
+
+
+class Request(object):
+    """Represents an incoming request to an endpoint.
+
+    Request class is used to represent the CallRequestMessage at User's level.
+    This is going to hide the protocol level message information.
+    """
+
+    # TODO decide which elements inside "message" object to expose to user.
+    def __init__(
+        self,
+        id=None,
+        flags=FlagsType.none,
+        ttl=10,
+        tracing=None,
+        service=None,
+        headers=None,
+        checksum=None,
+        argstreams=None,
+        scheme=None,
+    ):
+        self.flags = flags
+        self.ttl = ttl
+        self.service = service
+        self.tracing = tracing or Trace()
+        # argstreams is a list of InMemStream/PipeStream objects
+        self.argstreams = argstreams
+        self.checksum = checksum
+        self.id = id
+        self.headers = headers or {}
+        self.state = StreamState.init
+        self.endpoint = ""
+        self.header = None
+        self.body = None
+        self.scheme = scheme
+
+    @property
+    def arg_scheme(self):
+        return self.headers.get('as', None)
+
+    def set_exception(self, exception):
+        for stream in self.argstreams:
+            stream.set_exception(exception)
+            stream.close()
+
+    def close_argstreams(self, force=False):
+        for stream in self.argstreams:
+            if stream.auto_close or force:
+                stream.close()
+
+    @tornado.gen.coroutine
+    def get_header(self):
+        """Get the header value from the request.
+
+        :return: a future contains the deserialized value of header
+        """
+        raw_header = yield get_arg(self, 1)
+        if not self.scheme:
+            raise tornado.gen.Return(raw_header)
+        else:
+            header = self.scheme.deserialize_header(raw_header)
+            raise tornado.gen.Return(header)
+
+    @tornado.gen.coroutine
+    def get_body(self):
+        """Get the body value from the resquest.
+
+        :return: a future contains the deserialized value of body
+        """
+
+        raw_body = yield get_arg(self, 2)
+        if not self.scheme:
+            raise tornado.gen.Return(raw_body)
+        else:
+            body = self.scheme.deserialize_body(raw_body)
+            raise tornado.gen.Return(body)
+
+    def get_header_s(self):
+        """Get the raw stream of header.
+
+        :return: the argstream of header
+        """
+        return self.argstreams[1]
+
+    def get_body_s(self):
+        """Get the raw stream of body.
+
+        :return: the argstream of body
+        """
+        return self.argstreams[2]
