+++ python/tests/integration/json/test_json_server.py
@@ -25,8 +25,8 @@ import tornado
 import tornado.gen
 
 from tchannel.scheme import JsonArgScheme
-from tchannel.tornado import TChannel
 from tchannel.tornado import RequestDispatcher
+from tchannel.tornado import TChannel
 from tchannel.tornado.broker import ArgSchemeBroker
 from tests.integration.server_manager import TChannelServerManager
